Do not run Go tests in verbose mode
We have so many tests now in Crossplane that it's difficult to find which ones have failed amongst the wall of output when running `make test`. I don't think we're getting a lot of value from seeing every passing test name printed out; I hand hacked this to try it out and identifying which tests failed is a lot easier (and just as detailed) without `-v`.