reduce golangci-lint overhead for jenkins worker
The crossplane jenkins worker node only has 10gb of RAM.  With 1.17.1 of
golangci-lint we regularly see 10gb consumed while linting in local
environments.

The changes suggested here appear to improve the overhead.
https://github.com/golangci/golangci-lint/issues/337#issuecomment-510136513

Prior to the bump to 1.17.1, golangci-lint was taking 6-8GB of RAM and
completing in 10s.

With the reduction in overhead to golangci-lint that this PR introduces,
linting consumes between 8-9GB of ram and completes in 30s using the same
environment.

The intent here is to continue using latest golangci-lint with the
current Jenkins configuration, sacrificing some time.  Additional
measure can be made, such as using `--fast` tests only or further increasing the
LINT_GOGC value.

https://github.com/golangci/golangci-lint#configuration
https://github.com/golangci/golangci-lint#memory-usage-of-golangci-lint

Signed-off-by: Marques Johansson <marques@upbound.io>