helm: fix race condition for parallel package's
- Helm package save's chart to local chart repo by default. When
we have multiple charts in a repo and make is called with -j n,
parallel runs of helm package sporadically cause a race condition
while updating index.yaml file for local repo. We don't need this
functionality at all as we are already storing charts to our
output_dir and index later with a helm index target.

Signed-off-by: Hasan Turken <turkenh@gmail.com>