Build with Go 1.16.1, and remove the whitelist of supported versions
https://github.com/upbound/build/pull/122#discussion_r508284675

Per the above comment thread, most (all?) of our projects now use Go modules,
which allow a module to declare the version of Go it targets.

It could be that the purpose of this variable is to gate the versions of Go that
upbound/build as a library of Makefiles supports, rather than the versions of Go
that code built with upbound/build supports. If this is the case, I'd argue that
during the the two years I've been using this library we've yet to encounter a
new version of Go that has required anything other than a bump to this whitelist
and that we never test that the older versions of Go still work. I suspect it
would be more productive to pay the price of a newer version of Go breaking the
library if and when that becomes an issue than it currently is to bump the var
every time a new version of Go is released.

Signed-off-by: Nic Cope <negz@rk0n.org>