Optionally publish a valid docs version
Adds a flag to indicate that a certain version of the documentation
should not be published. If DOCS_VERSION_ACTIVE is set to false, the
version will be removed from documentation if it exists. This
effectively allows a given release branch to be retired by adding this
flag and merging the commit.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>