Use ARCH not TARGETARCH when selecting base image
We previously switched to using TARGETARCH when selecting the platform
for the base image used when building images in a repository. TARGETARCH
was added to work around building locally on Apple arm64 machines, and
the machinery that was implemented at that time caused TARGETARCH to be
derived from the HOSTARCH, even when a PLATFORM was specified for
building. This means that we now use an amd64 base image when the host
is amd64, even if the specified platform is arm64.

This reverts back to using ARCH, which is derived from PLATFORM. This
means that when linux_arm64 is supplied as platform, we will use arm64
as ARCH, rather than falling back to the host arch. This does not break
local builds for Apple arm64 machines because PLATFORM will be derived
from TARGETARCH since it is not explicitly provided.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>