Add imagelight as alternative image building orchestrator
Adds a new image building orchestration Makefile that is mutually
exclusive with image.mk. The purpose of imagelight.mk is to defer more
of the functionality to the individual image makefiles that exist in a
repo that consumes the build submodule. This use case has become more
desirable as we move towards tools such as buildx, but imagelight does
not specifically constrain the consumer to any one image building tool.

There are a few notable differences from image.mk, including but not
limited to:
- We do not save images as tarballs to an output directory, though
individual image makefiles in a repo could opt to do so on their own.
- We support publishing to multiple registries and repositories. An
arbitrary number of REGISTRY_ORGS can be defined in the top-level
Makefile in a repo, and each image defined in that repository will be
published to every specified location in REGISTRY_ORGS. REGISTRY_ORGS
take the format of registry-name/org-name (e.g. docker.io/crossplane).
- We skip building for platforms that are not valid, which, for now,
excludes all non-linux platforms. This simplifies the process of
specifying platforms in a repository because we do not force the
consumer to override platforms for which they want to build binaries but
do not want to or cannot publish images.
- We do not call common.mk targets when invoking the image-specific
Makefiles in a repository. This is an unnecessary level of indirection
that doesn't provide much additional value in any of the repos that
consume the build submodule. We instead call the targets in the image
Makefiles directly, requiring img.build, img.publish, and img.promote to
be defined.

Note that image.mk remain unmodified, which allows for any repositories
that currently consume it to continue doing so. I do not anticipate
removing image.mk any time soon as it provides different functionality
than imagelight.mk and some projects may opt never to switch.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>