Use /usr/bin/env to find bash
In some rare cases (e.g. on NixOS) bash is not at /bin/bash. Using env
to find where the shell actually is is a fairly common pattern per
https://en.wikipedia.org/wiki/Shebang_(Unix)#Portability. It's also
possible that /usr/bin/env is not at that path, but anecodally the
internet considers it to be "more likely" to be there than /bin/bash,
which is not the standard path on most BSDs and some rare Linux
variants.

Signed-off-by: Nic Cope <nicc@rk0n.org>